requirements.txt
flask==2.0.1
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==0.19.2
anthropic==0.5.0
gunicorn==20.1.0 